requirements.txt
firecrawl-py
phidata
pydantic
groq
python-dotenv
fastapi
uvicorn